Drawing current frame in viewer

+++ src/lib/modular_slam/orb_feature.cpp
@@ -27,12 +27,12 @@ std::vector<OrbKeypoint> OrbOpenCvDetector::detect(const RgbFrame& sensorData)
     descriptors.convertTo(descriptors, CV_32F);
 
     std::vector<OrbKeypoint> result;
-    keypoints.reserve(descriptors.rows);
+    keypoints.reserve(static_cast<std::size_t>(descriptors.rows));
 
     for(auto i = 0; i < descriptors.rows; ++i)
     {
         OrbKeypoint keypoint;
-        const auto& cvKeypoint = keypoints[i];
+        const auto& cvKeypoint = keypoints[static_cast<std::size_t>(i)];
         const auto* descriptor = descriptors.ptr<float>(i);
         keypoint.keypoint.coordinates = Eigen::Vector2f{cvKeypoint.pt.x, cvKeypoint.pt.y};
